rename $nn as $xxinv in atan_grind!() macro
I'm sure $nn made sense at the time, but it has little mnemonic
value now.  Renaming as $xxinv, because this is the (multiplicative)
inverse of the square of the "x" argument being passed to the Taylor
expansion.  In particular, if we are computing f(x), $xxinv = 1/(x*x).
Perhaps there is yet a better name for this value, but this is still
an improvment over $nn.