remove unnecessary type annotations
The compiler is perfectly capable of inferring the type of
some "let" expressions on its own; and adding explicit
type decorations does not necessarily aid the human reader.
Removing a couple of unecessary type annotations, on the
premise that they are just cluttering up the code without
adding any value.