improve code self-documentation of get_nwords() return-value computation
Give names to the "1" constants, so that the code self-documents better
(and remove the now-redundant comment).  Also name the div_ceil()
function we're rolling in place of the not-yet-std u32 method of the
same name.  Leave a stub comment as a reminder that some day we might
be able to rely on u32.div_ceil() being a thing, and not have to roll
our own.