Compute maxdigits, rather than having to guess and edit the source
The Theory.pdf file explains the math behind the equations being used, and
does an error analysis to derive how many digits can be computed without
courting integer overflows.  Make use of the latter to have the code
determine the maxdigits on its own, instead of requiring the editing of
a constant in the source code based on... a gut feeling?

This commit adds both Theory.tex (the source) and Theory.pdf (a pre-rendered
output of Theory.tex, on the assumption that most end-users would not have
a copy of pdflatex handy to do the rendering).  It also replaces what is
now redundant commentary in the source code with a pointer to the Theory
document.