Fix an error in the number-of-digits-limit derivation
A negative sign got lost in the derivation of the limit on $D$
(subtract-a-negative was left as a subtraction).  This patch fixes
that, and implications that fell out from that (e.g., instead of
having to subtract 20 from a simplified estimate, we can just drop
an additive term).  This also makes the statement about atan(1)
(about how few digits can be reliably computed) make more sense.