apply maxdigits computation correction
As noted in the recent Theory.tex commit, there was an
error in the derivation of the formula used to compute
the maximum number of digits that can be computed without
fear of integer overflows.  It does not have a meaningful
impact on the user interface side, but to keep the
conde consistent with the Theory, and to make a
trivial simplification to the code, I am dropping
the "- 20" term from the calculation.