Reset handling of SIGPIPE to SIG_DFL
Apparently rust sets SIGPIPE to SIG_IGN, and then println! calls will panic
if they see an EPIPE error.  This default behavior makes sense for, say,
a web server running some plug-in code, but we prefer the traditional
unix-y command-line behavior of just terminating.

There is a feature in nightly (tracking #97889) to be able to decorate fn
main() with #[unix_sigpipe = "sig_dfl"] to get the desired behavior, but
our code here is written to be run on stable rust.  So, as a work-around,
we make use of the (close to trivial) sigpipe crate instead.