Yet more re-factoring of core code
DRY: create divmod_step! macro as a unification of next_atan_numerator! and
next_atan_term().  Now divmod_step! must be a macro for two reasons: the old
"we insist on being able to see our divisor as a constant" reason (when that
is true), but also because when not true we want to take the type-distinct
but same-behavior libdivide::Divisor as the divisor.

Also, various renamings of variables and commentary changes were made as
related to the core change of this unification.